Different implementation to generate IAM db tokens.
Got rid of connection strings, using connection structures specific to db instead.

+++ src/app/cli/src/services/config/models.rs
@@ -556,7 +556,7 @@ pub struct RemoteDatabaseConfig {
     pub credentials_policy: DatabaseCredentialsPolicyConfig,
     pub database_name: String,
     pub host: String,
-    pub port: Option<u32>,
+    pub port: Option<u16>,
 }
 
 #[skip_serializing_none]
